fix: missing dependencies
- check if fonts already installed
- backup previous nvim configs

+++ nvim/install.sh
@@ -6,7 +6,9 @@ echo -e "\n=============================="
 echo "Installing dependencies..."
 echo "=============================="
 sudo apt update && sudo apt upgrade -y
-sudo apt install -y git wget curl unzip build-essential libfuse2
+sudo apt install -y git wget curl unzip build-essential libfuse2 python3
+PYTHON_VERSION=$(python3 -c "import sys; print(f'{sys.version_info.major}.{sys.version_info.minor}')")
+sudo apt install -y python3.${PYTHON_VERSION}-venv
 
 echo -e "\n=============================="
 echo "Installing NVM..."
@@ -29,11 +31,15 @@ echo "Installing JetBrains Mono font..."
 echo "=============================="
 FONT_DIR="$HOME/.fonts"
 mkdir -p "$FONT_DIR"
-wget -qO JetBrainsMono.zip "https://github.com/ryanoasis/nerd-fonts/releases/download/v3.3.0/JetBrainsMono.zip"
-unzip -o JetBrainsMono.zip -d "$FONT_DIR"
-rm JetBrainsMono.zip
-fc-cache -fv
-echo -e "\nJetBrains Mono font installed! Use it in your terminal."
+if [ ! -f "$FONT_DIR/JetBrainsMonoNerdFont-Regular.ttf" ]; then
+    wget -qO JetBrainsMono.zip "https://github.com/ryanoasis/nerd-fonts/releases/download/v3.3.0/JetBrainsMono.zip"
+    unzip -o JetBrainsMono.zip -d "$FONT_DIR"
+    rm JetBrainsMono.zip
+    fc-cache -fv
+    echo -e "\nJetBrains Mono font installed!"
+else
+    echo "JetBrains Mono font is already installed."
+fi
 
 echo -e "\n=============================="
 echo "Installing Neovim..."
@@ -44,9 +50,13 @@ sudo mv nvim-linux-x86_64.appimage /usr/local/bin/nvim
 echo -e "\nNeovim installed successfully!"
 
 echo -e "\n=============================="
-echo "Removing previous Neovim configurations..."
+echo "Backing up and removing previous Neovim configurations..."
 echo "=============================="
-rm -rf ~/.config/nvim ~/.local/state/nvim ~/.local/share/nvim
+
+BACKUP_DIR="$HOME/nvim_backup_$(date +%Y%m%d_%H%M%S)"
+mkdir -p "$BACKUP_DIR"
+mv ~/.config/nvim ~/.local/state/nvim ~/.local/share/nvim "$BACKUP_DIR/" 2>/dev/null
+echo "Backup created at: $BACKUP_DIR"
 
 echo -e "\n=============================="
 echo "Installing NvChad..."
@@ -61,4 +71,11 @@ mkdir -p ~/.config/nvim/lua/
 cp -r lua/custom ~/.config/nvim/lua/
 echo -e "\nCustom Neovim configuration applied!"
 
-echo -e "\nInstallation complete. Open Neovim by running: nvim\n"
+# reopen the terminal or ssh connection to apply the changes instead of running the following commands
+echo -e "\n=============================="
+export NVM_DIR="$HOME/.nvm"
+[ -s "$NVM_DIR/nvm.sh" ] && . "$NVM_DIR/nvm.sh"
+[ -s "$NVM_DIR/bash_completion" ] && . "$NVM_DIR/bash_completion"
+
+echo -e "\nIt is recommended to reopen the terminal or SSH connection to apply the changes."
+echo -e "\nInstallation complete! You can open Neovim by running: nvim\n"
